safehaven_local/safehaven-core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safehaven-core"
version = "0.1.0"
description = "SafeHaven v2 semantic state core"
requires-python = ">=3.10"
dependencies = [
  "PyYAML==6.0.2",
  "requests==2.32.3",
  "opencv-python-headless==4.10.0.84",
  "numpy==2.1.1",
  "prometheus-client==0.21.0",
  "paho-mqtt==2.1.0",
]

[project.scripts]
safehaven-core = "safehaven_core.main:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
